+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/A320_Neo_CDU_ATC_FlightReq.ts
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { FansMode } from '@datalink/common';
 import { CDUAtcClearanceReq } from './A320_Neo_CDU_ATC_ClearanceReq';
 import { CDUAtcMenu } from './A320_Neo_CDU_ATC_Menu';