+++ fbw-a380x/src/systems/instruments/src/PFD/LowerArea.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { Arinc429Values } from 'instruments/src/PFD/shared/ArincValueProvider';
 import {
   ClockEvents,